requirements.txt
# Web Application Framework
# 설치하세요
streamlit>=1.30.0

# Machine Learning & Deep Learning
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
datasets>=2.14.0
peft>=0.6.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0

# Vector Database & Embeddings
faiss-cpu>=1.7.4
huggingface-hub>=0.19.0

# Data Processing & Utilities
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.0

# Image Processing
Pillow>=10.0.0

# Web Scraping
selenium>=4.15.0
chromedriver-autoinstaller>=0.6.0

# Validation & Type Checking
pydantic>=2.0.0

# Standard Library Extensions
typing-extensions>=4.8.0 